client/requirements.txt
protobuf==3.20.3
PyYAML==6.0.1
urllib3==2.1.0
numpy==1.26.4
nvidia-pytriton==0.4.2
tritonclient==2.40.0
kafka-python==2.0.2
confluent-kafka==2.3.0
opencv-python==4.9.0.80
pycryptodome==3.20.0
pytz==2023.3.post1
requests==2.31.0
pillow==10.2.0
pymongo==4.6.2